% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Age.R
\docType{data}
\name{Age}
\alias{Age}
\title{Chapter 1 Age data}
\format{
A data frame with 10 rows and two columns (age and layoff)
}
\usage{
Age
}
\description{
This dataset come from a lawsuit by Robert Martin where he worked at Westavco, a paper producer company. He was laid off in Round 2 of several
layoffs by the company. He sued the company, claiming to be a victim of age discrimination. The ages of the 10 workers involved in Round 2 were: 25,
33, 35, 38, 48, 55, 55, 55, 56, and 64. The ages of the three people laid off were 55, 55, and 64.
}
\keyword{datasets}
