% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3.R
\docType{data}
\name{AIDS}
\alias{AIDS}
\title{Chapter 3 AIDS Data}
\format{
A data frame with 41 rows and 4 columns
}
\usage{
AIDS
}
\description{
The AIDS data set shows the prevalence of HIV/AIDS in adults aged 15 to 49 in least developed
countries along with country-level data on several socioeconomic factors. Use this data set to create a
multivariate regression model to predict the prevalence of HIV/AIDS from socioeconomic factors.
}
\keyword{datasets}
