% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C8.R
\docType{data}
\name{APExams}
\alias{APExams}
\title{Chapter 8 APExams}
\format{
A data frame with 100 rows and three columns
}
\usage{
APExams
}
\description{
Are high-achieving high school students more inclined to take Advanced Placement exams? In a
recent student survey, the co-author of this book asked students in his statistics courses to report the
number of Advanced Placement (AP) exams they had taken in high school. In addition, information
on their self-reported cumulative high school grade point average (HSGPA) and gender was recorded.
Assume that students attended high school for four years.
}
\keyword{datasets}
