% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3.R
\docType{data}
\name{Arsenic}
\alias{Arsenic}
\title{Chapter 3 Arsenic Data}
\format{
A data frame with 21 rows and 6 Columns
}
\source{
M. Karagas, J. Morris, J. Weiss, V. Spate, C. Baskett, and E. Greenberg, “Toenail Samples as an Indicator of Drinking Water Arsenic Exposure,” Cancer Epidemiology, Biomarkers and Prevention, 5 (1996): 849–852
}
\usage{
Arsenic
}
\description{
Karagas, Morris, Weiss, Spate, Baskett, and Greenberg collected toenail clippings of several individuals along with samples of each person’s regular drinking water.6
In this pilot study, the researchers attempted to determine if arsenic concentrations in people’s toenails could be used to determine
the level of arsenic in their water. Each of the 21 individuals in this study had a private well. High
arsenic concentrations are related to cancer, and several studies have found a positive correlation
between arsenic concentrations of drinking water and toenail samples.
}
\details{
DRINK: Household well used for drinking (1=<1/4 2=1/4 3=1/2 4=3/4 5=>3/4).
COOK: Household well used for cooking (1=<1/4 2=1/4 3=1/2 4=3/4 5=>3/4).
WATER: Arsenic in water (ppm).
NAILS: Arsenic in toenails (ppm).
}
\keyword{datasets}
