% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C6.R
\docType{data}
\name{Baboon}
\alias{Baboon}
\title{Chapter 6 Baboon}
\format{
A data frame with 242 rows and nine columns
}
\source{
V. K. Bentley-Condit, T. L. Moore, and E. O. Smith, “Analysis of Infant Handling and the Effects of Female Rank Among Tana River Adult Female Yellow Baboons Using Permutation/Randomization Tests,” American Journal of Primatology, 55 (2001): 117–130
}
\usage{
Baboon
}
\description{
Primatologist Vicki Bentley-Condit of Grinnell College studied interactions between female and infant
yellow baboons (Papio cynocephalus cynocephalus) at the Tana River National Primate Reserve in Kenya.
She collected the data for her study by observing baboons in 20-minute focal samples over an 11-month
period in 1991–1992. Her subjects included 23 female baboons, 11 of which were mothers with infants (no
mother with more than one offspring). Bentley-Condit observed and recorded interactions between females
and infants, excluding interactions between a mother and her own offspring.
}
\keyword{datasets}
