% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C6.R
\docType{data}
\name{Baboon_Variable_Descriptions}
\alias{Baboon_Variable_Descriptions}
\title{Chapter 6 Baboon_Variable_Descriptions}
\format{
A data frame with nine rows and four columns
}
\usage{
Baboon_Variable_Descriptions
}
\description{
A codebook for the Baboon data.
}
\keyword{datasets}
