% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C6.R
\docType{data}
\name{BaboonVariableDescriptions}
\alias{BaboonVariableDescriptions}
\title{Chapter 6 BaboonVariableDescriptions}
\format{
A data frame with nine rows and four columns
}
\usage{
BaboonVariableDescriptions
}
\description{
A codebook for the Baboon data.
}
\keyword{datasets}
