% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C6.R
\docType{data}
\name{babyweights}
\alias{babyweights}
\title{Chapter 6 babyweights}
\format{
A data frame with 135 rows and one column
}
\usage{
babyweights
}
\description{
The file BabyWeight provides newborn weights for a simple random sample of 135 infants born in
the United States in 1995.
}
\keyword{datasets}
