% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C4.R
\docType{data}
\name{Bacteria}
\alias{Bacteria}
\title{Chapter 4 Bacteria}
\format{
A data frame with 36 rows and three columns (Building, Location, and Count)
}
\usage{
Bacteria
}
\description{
Two students, Isaac and Courtney, sampled surfaces around their campus to analyze the prevalence of bacteria.
They compared different types of surfaces in both residential and academic buildings.
}
\details{
Data were collected by wiping surfaces with a wet Q-tip and swabbing the result on a standard
nutrient agar plate. Locations were all tested on April 24, 2009 over the course of a two-hour period.
The plates were incubated at 37°C for 48 hours before colony-forming units (CFUs) were counted as a
measure of bacteria levels. Figure 4.5 shows several of the plates from this study. When CFUs exceeded
400 per plate, one fourth of the plate was counted and the total was calculated from that sample.

Six different buildings were swabbed, with two buildings representing each type of facility:
• Academic buildings: Noyce and ARH
• Public residential buildings: Norris and Dibble
• Private residences: Cowles apartment 7110 and Yellow House (1478 Park St.)
Within each building, two faucets, two door handles, and two desk surfaces were swabbed.
}
\keyword{datasets}
