% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C7.R
\docType{data}
\name{Birdkeeping}
\alias{Birdkeeping}
\title{Chapter 7 Birdkeeping Data}
\format{
A dataset with 147 rows and 7 columns
}
\source{
P. A. Holst, D. Kromhout, and R. Brand, “For Debate: Pet Birds as an Independent Risk Factor for Lung Cancer,” British Medical Journal, 297.6659 (1988): 6659 1319–1321.
}
\usage{
Birdkeeping
}
\description{
Holst, Kromhout, and Brand conducted a case-control study to determine if keeping birds
increased the chances of getting lung cancer. It is believed that people who keep birds may
inhale more allergens and dust particles and thus may be more likely to get lung cancer. They collected data on 49 people with lung cancer and 98 people with similar demographics who did not
have lung cancer.
}
\keyword{datasets}
