% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C7.R
\docType{data}
\name{Burns}
\alias{Burns}
\title{Chapter 7 Burns Data}
\format{
A dataset with 435 rows and two columns
}
\source{
J. Fan, N. E. Heckman, and M. P. Wand, “Local Polynomial Kernel Regression for Generalised Linear Models and Quasi-Likelihood Functions,” Journal of the American Statistical Association, 90 (1995): 47 141–150.
}
\usage{
Burns
}
\description{
Fan, Heckman, and Wand analyzed third-degree burn data from the University of Southern California General Hospital Burn Center. In the Burns data set, 435 patients (adults ages 18–85) were
grouped according to the size of the third-degree burns on their body. The explanatory variable is
listed as the midpoint of set intervals: ln (area in square centimeters +1). The response in this data
set is whether or not the patient survived (1 represents a survival).
}
\keyword{datasets}
