% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C7.R
\docType{data}
\name{Cancercells}
\alias{Cancercells}
\title{Chapter 7 Cancercells}
\format{
A dataset with 569 rows and 2 columns
}
\usage{
Cancercells
}
\description{
Cancercells shows the values of the median radius and sees whether they are malignant or not
}
\keyword{datasets}
