% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C8.R
\docType{data}
\name{CancerCluster}
\alias{CancerCluster}
\title{Chapter 8 CancerCluster}
\format{
A data frame with eight rows and six columns
}
\usage{
CancerCluster
}
\description{
Sample Cancer Data
}
\keyword{datasets}
