% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Car1.R
\docType{data}
\name{Car1}
\alias{Car1}
\title{Chapter 1 Car1 dataset}
\format{
A data frame with 41 rows and 4 columns (Price, Make, Sound, Leather)
}
\usage{
Car1
}
\description{
The Car1 data set contains prices of used 2005 General Motors cars.
All cars in this sample are six-cylinder, four-door sedans in excellent condition with premium sound systems
and leather interiors. The data set contains four makes of General Motors cars: Pontiac, Chevrolet, Cadillac,
and Buick.
}
\section{Chapter 1}{
NA
}

\keyword{datasets}
