% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3.R
\docType{data}
\name{Cars}
\alias{Cars}
\title{Chapter 3 Cars Data}
\format{
A data frame with 804 rows and 12 Columns
}
\usage{
Cars
}
\description{
The data set Cars contains the make, model, equipment, mileage, and Kelley Blue Book
suggested retail price of several used 2005 GM cars. For more information, see page 68.
}
\keyword{datasets}
