% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C10.R
\docType{data}
\name{CarsC10}
\alias{CarsC10}
\title{Chapter 10 CarsC10}
\format{
A data frame with 804 rows and 19 columns
}
\usage{
CarsC10
}
\description{
The data set Cars contains the make, model, equipment, mileage, and Kelley Blue Book suggested
retail price of several used 2005 General Motor cars. Kelley Blue Book (http://www.kbb.com) has been
an accurate resource for pricing cars for over 80 years and was used as a source to collect these data.
}
\keyword{datasets}
