% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3.R
\docType{data}
\name{Caucuses}
\alias{Caucuses}
\title{Chapter 3 Caucuses Data}
\format{
A data frame with 99 rows and 11 columns
}
\usage{
Caucuses
}
\description{
The Caucuses data set contains the 2008 Democratic and Republican caucus results. Republicans
count actual votes. Democrats don’t record individual votes, but each precinct allocates its delegates
based on the number of people in the precinct supporting each candidate during the caucus. Senator
Clinton was expected to win in most polls, but ended up a surprising third after Senators Obama and
Edwards. Many political analysts have found that females, less educated people, and older people
voted for Senator Clinton, while younger people, more educated people, and African Americans preferred Senator Obama.
}
\keyword{datasets}
