% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3.R
\docType{data}
\name{Cavalier}
\alias{Cavalier}
\title{Chapter 3 Cavalier Data}
\format{
A data frame with 30 rows and 12 Columns
}
\usage{
Cavalier
}
\description{
The data set is a subset of the Cars dataset that only contains Cavalier models. See ?Cars for more information.
}
\keyword{datasets}
