% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChiSq.R
\docType{data}
\name{ChiSq}
\alias{ChiSq}
\title{Chapter 1 ChiSq data}
\format{
A data frame with 5000 rows and one column
}
\usage{
ChiSq
}
\description{
The ChiSq data comes from a highly skewed population (with mean 0.9744 and standard deviation 1.3153)
}
\keyword{datasets}
