% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C5.R
\docType{data}
\name{Colors}
\alias{Colors}
\title{Chapter 5 Colors data}
\format{
A data frame with 128 rows and 5 columns
}
\usage{
Colors
}
\description{
Kastner et al. studied how the brain recognizes the shape and color of an object.
They found that the
process of identifying the shape and color of an item is carried out not simultaneously but in steps.
Two students decided to investigate the impact of color distraction on a shape matching game called
Shapesplosion. In this game, subjects are asked to match shapes by placing specifically shaped pegs
into the matching hole as quickly as possible. These student researchers selected eight female science majors and eight female non-science majors at their college. Each of the 16 subjects played four
games (each of the following four games in random order):
Factor 1 (game complexity): the Shapesplosion game with 15 pieces (pegs) or the game with 18 pieces
Factor 2 (color distracter): the Shapesplosion game where the peg color matches the hole color or the
game where the pegs are a different color than the hole
The Time it took (in seconds) to complete each game was recorded.
}
\keyword{datasets}
