% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C5.R
\docType{data}
\name{Colors2}
\alias{Colors2}
\title{Chapter 5 Colors2 data}
\format{
A data frame with 248 rows and 6 columns
}
\usage{
Colors2
}
\description{
The design in Exercise 11 is sometimes called a split-plot/repeated measures (SP/RM) \link{1,2}, where
the 1 represents the number of whole-plot factors and the 2 represents the number of split-plot
factors.
In fact, the previous exercise provides only half of the data. In addition to the 16 females discussed
in Exercise 11, 16 males were also tested. The complete design is a split-plot/repeated measures (SP/RM) \link{2,2}, with both gender and division as whole-plot factors.
}
\keyword{datasets}
