% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C7.R
\docType{data}
\name{Convict}
\alias{Convict}
\title{Chapter 7 Convict Data}
\format{
A dataset with 60 rows and two columns.
}
\source{
S. Wilson and B. Mandelbrote, “Drug Rehabilitation and Criminality,” British Journal of Criminology, 18 (1978): 381–386.
}
\usage{
Convict
}
\description{
A study was conducted to determine if a relationship existed between criminal conviction and years
of education. Sixty people who had taken part in a drug rehabilitation program were classified by
years of education. Each person was categorized as having a “short” education (15 years or less) or a “long” education (more than 15 years); also recorded was whether or not they had a post-treatment
conviction.
}
\keyword{datasets}
