% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C5.R
\docType{data}
\name{Cookies}
\alias{Cookies}
\title{Chapter 5 Cookies Data}
\format{
A data frame with 135 rows and 4 columns
}
\usage{
Cookies
}
\description{
Two students wanted to determine if people could taste the difference in chocolate chip cookies with
varying amounts of sugar and varying amount of freshness. Nine batches were made, following the
recipe on the chocolate chip bag as closely as possible except for the amount of sugar. Each batch
was randomly assigned to one of three treatments: half the suggested amount of sugar, the suggested
amount of sugar, and double the suggested amount of sugar.
On the day the nine batches of cookies were baked, the researchers handed out five cookies
from each batch (a total of 45 cookies) to people in their dorm and asked them to rate the cookies from 1 through 10, with 1 being inedible and 10 being the best cookie they every had. The researchers stored the rest of the cookies for a day. On the second day, the researchers handed out five more
cookies from each of the original nine batches to students in their dorm and asked them to rate them
from 1 through 10. The researchers did the same thing on the third day.
}
\keyword{datasets}
