% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C5.R
\docType{data}
\name{Corn}
\alias{Corn}
\title{Chapter 5 Corn Data}
\format{
A data frame with 40 rows and 5 columns
}
\source{
A. Reza, Design of Experiments for Agriculture and the Natural Sciences, 2nd ed. (Boca Raton, FL: Chapman & Hall CRC, 2006), p. 138.
}
\usage{
Corn
}
\description{
A study was conducted to determine if different species of corn or amounts of nitrogen would
impact yields.
Nitrogen is available in the soil as the result of natural and biological processes;
however, farmers often add a nitrogen fertilizer to corn crops in order to increase yields. A nitrogen
deficiency can considerably decrease yields. Fertilizer is expensive; thus, many farmers will choose
to plant a species of corn that grows well with less nitrogen. Soil types vary with each plot of land.
The amount of nitrogen can also vary from year to year, depending on the type of crop (e.g., wheat, oats, hay, soybeans, or corn) planted the previous year. Thus, the amount of nitrogen needed will be
different for different plots. Researchers often conduct studies for a county or entire state, so they
are primarily interested in measuring the variability between plots instead of how well a specific plot
does compared to another. Determining the right species to plant as well as the right amount of fertilizer can dramatically impact a farmer’s profit. In fact, a quick Google search will show that many
agricultural states provide corn-nitrogen productivity calculators on the web.
Many seed companies and state-sponsored agricultural research centers have large testing areas
that have been shown to be similar to other plots of land throughout their state. In a large testing
area, eight 20-acre plots of land were randomly assigned to a nitrogen application rate (0, 70, 140, or 210 pounds of nitrogen per acre). Each of these 20-acre plots was also subdivided into five 4-acre
subplots. Within each 20-acre plot, the subplots were assigned to be planted with one species of
hybrid corn (A, B, C, D, or E). At the end of the season, the fields were harvested and the yield in
bushels per acre was recorded.
}
\keyword{datasets}
