% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C10.R
\docType{data}
\name{CrimeC10}
\alias{CrimeC10}
\title{Chapter 10 CrimeC10}
\format{
A data frame with 47 rows and eight columns
}
\usage{
CrimeC10
}
\description{
The FBI’s Uniform Crime Reporting (UCR) program gathers crime data from more than 17,000 law
enforcement agencies throughout the United States. The Crime data set contains crime counts for
several years in Iowa. Crimes are listed by type of crime (murder, burglary, larceny, etc.).
}
\keyword{datasets}
