% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C4.R
\docType{data}
\name{Cups}
\alias{Cups}
\title{Chapter 4 Cups}
\format{
A data frame with 32 rows and seven columns
}
\usage{
Cups
}
\description{
A manager of a manufacturing company, shift managers, cup machine operators,
and a lone statistician decided to identify which factors were most influential in keeping their cups from leaking.
Over 30 possible factors were identified, but after some thoughtful discussions the group settled on
six variables that should be tested for their effects on leaking cups. One of the six factors of interest
was which paper supplier to use. Since the company was considering changing suppliers, funds were
available to do some product testing before a purchase was made. However, each trial (each run of
production under specified factor conditions) would cost the company thousands of dollars in lost
production time, material costs, and employee costs. T
}
\details{
The company agreed to conduct 32 tests, but wanted to test all six factors and
all corresponding twoway interactions. Fractional factorial designs are very useful
for this type of exploratory data analysis. The details of fractional factorial designs
are beyond the scope of this text. However, balanced
data are a key concept behind these designs. For example, in the Cups data, every factor has two levels
and each level has 16 observations. In addition, within the first factor level (the 16 observations
where side-seam temperature is set to 70\%), every other factor is still balanced (every other factor
has 8 observations at each level).
}
\keyword{datasets}
