% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C7.R
\docType{data}
\name{Donner}
\alias{Donner}
\title{Chapter 7 Donner}
\format{
A datasets with 87 rows and 5 columns
}
\source{
D. K. Grayson, “Donner Party Deaths: A Demographic Assessment,” Journal of Anthropological Research, 46, 3 (1990). Also see http://www.xmission.com/~octa/DonnerParty/Roster.html.
}
\usage{
Donner
}
\description{
In 1846, a group of 87 people (called the Donner Party) were heading west from Springfield, Illinois,
for California.16 The leaders attempted a new route through the Sierra Nevada and were stranded
there throughout the winter. The harsh weather conditions and lack of food resulted in the death of
many people within the group. Social scientists have used the data to study the theory that females
are better able than males to survive harsh conditions.
}
\keyword{datasets}
