% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C2.R
\docType{data}
\name{Emission}
\alias{Emission}
\title{Chapter 2 Emission}
\format{
A data frame with 78 rows and two columns (Year and Emission)
}
\usage{
Emission
}
\description{
The data set Emission provides hydrocarbon emission in parts per million (ppm)
at idling speed for cars, based on the year each car was manufactured. These
data were randomly sampled froom a much larger study on pollution control in
Albuquerque, New Mexico.
}
\keyword{datasets}
