% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C9.R
\docType{data}
\name{Firstdrink}
\alias{Firstdrink}
\title{Chapter 9 Firstdrink}
\format{
A dataset with 1000 rows and 3 columns
}
\usage{
Firstdrink
}
\description{
When do individuals have their first drink of alcohol? The legal age for consuming alcohol is 21 years, but
some individuals claim to have had their first alcoholic drink when they were as young as 1 year old! We can
investigate the age at which individuals had their first drink of alcohol using data from the National Comorbidity Survey of 1990–1992. Participants were asked to recall the age at which they had their first drink of
alcohol. Those who could recall the age at which they had their first drink had complete event times (i.e., the                                                                                                      age at which they had their first drink). Individuals who had not had a drink by the time of the interview had
a right-censored event time (their age at the time of the interview). Individuals who could not recall the age
at which they had their first drink were not included in the sample.
}
\keyword{datasets}
