% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C9.R
\docType{data}
\name{Fruitfly}
\alias{Fruitfly}
\title{Chapter 9 Fruitfly}
\format{
A dataset with 125 rows and 6 columns
}
\source{
L. Partridge and M. Farquhar, “Sexual Activity and the Lifespan of Male Fruitflies,” Nature, 294 (1981): 580–581;
J. A. Hanley, “Appropriate Uses of Multivariate Analysis,” Annual Review of Public Health, 4 (1983): 155–180; and
J. A. Hanley and S. H. Shapiro, “Sexual Activity and the Lifespan of Male Fruitflies: A Data set That Gets Attention,”
Journal of Statistics Education, 2 (1994).
}
\usage{
Fruitfly
}
\description{
One misconception that some students initially have about survival analysis methods is that they can
be applied only to survival data that contain some censored observations. While survival analysis
methods are appropriate for incomplete data, they are also perfectly acceptable for noncensored survival data. Earlier we noted that the empirical survival function and Kaplan-Meier estimator are identical when there are no censored event times.
The data set Fruitfly, introduced by Partridge and Farquhar and further analyzed by Hanley
and Hanley and Shapiro, was originally analyzed for the purpose of investigating the relationship
between increased sexual activity of male fruitflies and longevity of life (in days) using regression and
analysis of covariance techniques. However, survival analysis methods can also be used to study the
life durations of male and female fruit flies. See pages 325 and 326 for brief descriptions of the variables.
}
\details{
Partners: number of companions (0, 1, or 8)
Type: type of companion (0 = newly pregnant female, 1 = virgin female, 9 = not applicable (when                                                                                        Partners = 0))
Longevity: lifespan, in days (This is the time-to-event variable.)
Thorax: length of thorax in mm
Sleep: percentage of each day spent sleeping
Censor: censoring status (Note that this variable takes only value 1, since the data are all complete. A
censoring status variable is necessary for software implementation.)
}
\keyword{datasets}
