% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C8.R
\docType{data}
\name{Gala}
\alias{Gala}
\title{Chapter 8 Gala}
\format{
A data frame with 30 rows and eight columns
}
\usage{
Gala
}
\description{
Data on plant species on the Galápagos islands have been collected more than once. The most recent
extensive data were reported by Johnson and Raven. We will consider two response variables: island total
observed species count and island endemic species count. There are five potential covariates that we expect
to be related to species count: island area (km2
) and elevation (meters), the distance (km) from the island to
its nearest neighbor (adjacent island) and to the largest island (km2
) in the archipelago (Santa Cruz), and the
area of the adjacent island (km2).
}
\keyword{datasets}
