% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C2.R, R/C4.R
\docType{data}
\name{Games2}
\alias{Games2}
\title{Chapter 2 Games2}
\format{
A data frame with 40 rows and six columns (studentID, numpegs, Type, Time, Hand, and Type2).

A data frame with 40 rows and six columns
}
\usage{
Games2

Games2
}
\description{
An extension on the Games1 dataset

The data set Games2 shows a column Type2 with four types of games based on distracter and
which hand was used.
}
\keyword{datasets}
