% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C4.R
\docType{data}
\name{Games2C4}
\alias{Games2C4}
\title{Chapter 4 Games2C4}
\format{
A data frame with 40 rows and six columns
}
\usage{
Games2C4
}
\description{
The data set Games2 shows a column Type2 with four types of games based on distracter and
which hand was used.
}
\keyword{datasets}
