% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C10.R
\docType{data}
\name{Grades}
\alias{Grades}
\title{Chapter 10 Grades}
\format{
A data frame with 32 rows and 14 columns
}
\usage{
Grades
}
\description{
The data set Grades contains total scores for students’ homework, labs, quizzes, projects, and exams. This course had
two 1-hour exams and a 3-hour final exam. In addition to a final project worth 20\% of the final grade, there
was an initial small project worth only 5\% of the final grade. The exams were fairly standard statistics exams,
but both projects required students to read peer-reviewed research projects, collect their own data, conduct an
appropriate statistical analysis, and present their results. While several students in this introductory class had
seen a little statistics in previous courses, this was typically the first time students were asked to read journal
articles or conduct a realistic research project.
}
\keyword{datasets}
