% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C10.R
\docType{data}
\name{Grades_details}
\alias{Grades_details}
\title{Chapter 10 Grades_details}
\format{
A data frame with 32 rows and 11 columns
}
\usage{
Grades_details
}
\description{
Data supplementary to the GradesC10 data set.
}
\keyword{datasets}
