% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C9.R
\docType{data}
\name{Graduate}
\alias{Graduate}
\title{Chapter 9 Graduate}
\format{
a dataset with 1000 rows and 3 columns
}
\usage{
Graduate
}
\description{
We can investigate the number of years needed to complete the requirements for a bachelor’s degree using data from individuals who participated in the National Educational
Longitudinal Survey (NELS) from 1988 to 2002. To illustrate how survival analysis techniques can be
used to answer these questions, we will use data on a sample of 1000 participants from the NELS data set
who began college prior to the year 2000. Note that this sample includes individuals who began study at
any community college or bachelor’s degree granting postsecondary institution, even though some students
who began their postsecondary education at a two-year college may not have been intending to pursue a
bachelor’s degree.
The time-to-event random variable is the number of years taken to complete the requirements for a bachelor’s degree. If an individual had obtained a bachelor’s degree by the interview date, then her or his event
time is complete. If the student had dropped out or had not graduated by the interview date, then her or his
time is right censored.
}
\keyword{datsets}
