% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C8.R
\docType{data}
\name{Grandslam}
\alias{Grandslam}
\title{Chapter 8 Grandslam}
\format{
A data frame with 30 rows and 21 columns
}
\usage{
Grandslam
}
\description{
The data set Grandslam contains measurements on several commonly recorded
statistics (e.g., numbers of grand slams, home runs, and runs batted) for 30 Major League Baseball teams. The
values for the variables GrSlam (number of grand slams), Cycles (number of cycles), and Games (number
of games) are count totals over five seasons from 2006 to 2010. The values for the remaining variables are
averages over the five seasons.
}
\keyword{datasets}
