% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C5.R
\docType{data}
\name{Handwash}
\alias{Handwash}
\title{Chapter 5 Handwash Data}
\format{
A data frame with 12 rows and 5 columns
}
\usage{
Handwash
}
\description{
With the H1N1 outbreak in 2010, hand washing was well promoted as one of the most effective and
economical methods for preventing the spread of infectious diseases. For a final project, two statistics students
were interested in comparing the effectiveness of three hand-washing soaps (62\% ethyl-alcohol-based hand
Sanitizer, 0.15\% tricolsan antibacterial soap, and a soap lacking antimicrobial agents).
These researchers used sterile swabs to swab each participant’s right hand before washing, first going
around the fingers and then making an S-shape on the palm. The right hand was swabbed again after subjects
washed their hands with one of the cleansers for 20 seconds and let their hands air dry for 3 minutes.
Swabs were then placed into a microfuge tube with 500 mL of saline and swirled for 30 seconds to knock
off bacteria. Then the researchers vortexed the microfuge tube for 20 seconds and pipetted 100 mL of bacterial solution onto an L-agar plate. Plates were then put in an incubator at 30°C for 72 hours. Colony-forming
units (CFUs) were then counted on each plate.
}
\keyword{datsets}
