% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C2.R
\docType{data}
\name{Hodgkins}
\alias{Hodgkins}
\title{Chapter 2 Hodgkins}
\format{
A data frame with 65 rows and two columns (PatientType and PlasmaLevels).
}
\usage{
Hodgkins
}
\description{
The data set Hodgkins contains plasma bradykininogen levels (in micrograms of bradykininogen
per milliliter of plasma) in three types of subjects (normal, patients with active Hodgkin’s disease,
and patients with inactive Hodgkin’s disease). The globulin bradykininogen is the precursor substance
for bradykinin, which is thought to be a chemical mediator of inflammation.
}
\keyword{datasets}
