% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C9.R
\docType{data}
\name{Hornhonk}
\alias{Hornhonk}
\title{Chapter 9 Hornhonk}
\format{
A dataset with 57 rows and 2 columns
}
\source{
A. Diekmann, M. Jungbauer-Gans, H. Krassnig, and S. Lorenz, “Social Status and Aggression: A Field Study
Analyzed by Survival Analysis,” Journal of Social Psychology, 136 (1996): 761–768.
}
\usage{
Hornhonk
}
\description{
In a study on aggressive behavior displayed by motorists, Diekmann and his colleagues investigated the time it took for 57
motorists intentionally blocked at a green light by a Volkswagen Jetta to show signs of aggression. Signs of
aggression included honking their horn or beaming their headlights at the Jetta. The time-to-event variable is
the time (measured in seconds to the nearest hundredth of a second) until the motorist honked (or beamed the headlights) at the Jetta.
If the motorist did not honk or flash the headlights by the time the Jetta moved, then the motorist’s event time was right censored.
}
\keyword{datasets}
