% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C8.R
\docType{data}
\name{Hurricane}
\alias{Hurricane}
\title{Chapter 8 Hurricane}
\format{
A data frame with 59 rows and six columns
}
\usage{
Hurricane
}
\description{
With all the discussion about climate change in the past couple of decades, it is no surprise that the
occurrence of several intense hurricanes over the past several years has garnered some attention. You
may remember that in 2005 the U.S. Gulf coast was hit by hurricanes Emily, Katrina, and Rita. All three
were category five on the Saffir-Simpson scale. That is the highest intensity on this five-level scale.
}
\details{
The Data set hurricanes contains 59 lines, one for each year from 1950 through 2008. For each
year, there are five columns of data containing counts of the number of hurricanes of each category
type on the Saffir-Simpson scale (labeled cat1, cat2, cat3, cat4, or cat5). The data were
obtained from http://weather.unisys.com/hurricane/atlantic
}
\keyword{datasets}
