% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C10.R
\docType{data}
\name{Intelligence}
\alias{Intelligence}
\title{Chapter 10 Intelligence}
\format{
A data frame with 214 rows and 10 columns
}
\usage{
Intelligence
}
\description{
The data set Intelligence contains test results for 214 people who have each taken several
intelligence tests.
}
\keyword{datasets}
