% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C2.R
\docType{data}
\name{Jury}
\alias{Jury}
\title{Chapter 2 Jury}
\format{
A data frame with 46 rows and two columns (Judge and Percent Women).

A data frame with 46 rows and two columns (Judge and Percent Women).
}
\usage{
Jury

Jury
}
\description{
Dr. Benjamin Spock was a well-known pediatrician who faced trial in 1968 for his activities as a
Vietnam War protester. Specifically, he was charged with conspiring to violate the Selective Service Act by encouraging young men to resist the draft. As part of his defense, his counsel claimed
that women were underrepresented on the jury. Women tend to be more sympathetic toward war
protesters than men do. The defense counsel claimed that the judge had a history of choosing
juries on which women were systematically underrepresented. At that time, jury members in Boston were chosen from a venire (a group of 30 to 200 individuals preselected from the population
by the judge’s clerk). By law, people were supposed to be selected for a venire at random. For Dr.Spock’s trial, the judge’s venire had 100 people and only 9 women, none of whom were selected
to be on the actual jury.
Dr. Spock’s defense counsel collected data on the percentages of women in venires from this
judge’s recent trials together with those of other judges in the Boston area.

Dr. Benjamin Spock was a well-known pediatrician who faced trial in 1968 for his activities as a
Vietnam War protester. Specifically, he was charged with conspiring to violate the Selective Service Act by encouraging young men to resist the draft. As part of his defense, his counsel claimed
that women were underrepresented on the jury. Women tend to be more sympathetic toward war
protesters than men do. The defense counsel claimed that the judge had a history of choosing
juries on which women were systematically underrepresented. At that time, jury members in Boston were chosen from a venire (a group of 30 to 200 individuals preselected from the population
by the judge’s clerk). By law, people were supposed to be selected for a venire at random. For Dr.Spock’s trial, the judge’s venire had 100 people and only 9 women, none of whom were selected
to be on the actual jury.
Dr. Spock’s defense counsel collected data on the percentages of women in venires from this
judge’s recent trials together with those of other judges in the Boston area.
}
\keyword{datasets}
