% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C1.R
\docType{data}
\name{MedSalaries}
\alias{MedSalaries}
\title{Chapter 1 MedSalaries data}
\format{
A data frame with 100 rows
}
\usage{
MedSalaries
}
\description{
The file MedSalaries is a random sample
of n = 100 salaries of medical doctors who were teaching at United States universities in 2009. For more information,
refer to page 16.
}
\keyword{datasets}
