% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C9.R
\docType{data}
\name{MeltingChipsJS}
\alias{MeltingChipsJS}
\title{Chapter 9 Melting Chips}
\format{
A dataset with 22 rows and 3 columns
}
\usage{
MeltingChipsJS
}
\description{
This dataset comes from a statistics class of 22 students where half were given a white chocolate chip and the other
a milk chocolate chip. They were then told to place the chocolate chip in their mouth and record the time it
takes for it to completely melt. The censor variable says that either something went wrong or the chip had
not melted till the cutoff point (which in this case is 75 seconds).
}
\keyword{datasets}
