% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C5.R
\docType{data}
\name{Memory}
\alias{Memory}
\title{Chapter 5 Memory Data}
\format{
A dataframe with 48 rows and 8 columns
}
\usage{
Memory
}
\description{
This is the main dataset used in chapter 5. In the memory study, 12 subjects each took 4 tests, providing 48 measurements. This may not appear to be
a large enough sample size to simultaneously test the three research questions for this study. Statistics based
on larger sample sizes are more reliable and increase the ability to confidently draw conclusions. However,
properly designed experiments can be very effective even with small sample sizes, as we will see when we
analyze the memory experiment data. Instead of attempting to calculate the sample size needed before a study
is conducted, this text focuses on using the model assumptions and analysis (discussed in Chapters 2 and 4)
to determine if the study is reliable. See chapter 5 for more information.
}
\keyword{datasets}
