% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C4.R
\docType{data}
\name{MemoryA}
\alias{MemoryA}
\title{Chapter 4 MemoryA}
\format{
A data frame with 100 rows and three columns
}
\usage{
MemoryA
}
\description{
Michael Eysenck tested 100 subjects (50 people between the ages of 55 and 65, 50 younger people)
to determine if there was a relationship between age and memory.10 Each subject was shown 27
words and asked to recall as many of those words as possible. He also tested whether five different
techniques impacted memory. Each subject was given one of five types of instructions:
Counting: count the number of letters in each word
Rhyming: think of a word that rhymes with each word
Adjective: think of an adjective to describe each word
Imagery: create an image of each word
Intentional: remember as many words as possible
}
\details{
The subjects in the first four groups were not aware that they would later be asked to recall each
word.
}
\keyword{datasets}
