% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C1.R
\docType{data}
\name{Mice}
\alias{Mice}
\title{Chapter 1 Mice data}
\format{
A data frame with 4 columns and 5 rows
}
\usage{
Mice
}
\description{
This data comes from a study in which the researchers wanted to find out whether K11777 helps stop schistosome worms from growing.
In one phase of the study, 10 female laboratory mice and 10 male laboratory mice were deliberately infected with the schistosome parasite.
Starting seven days after being infected with schistosomiasis, each mouse was given an injection every day for 28 days. Within each sex,
five mice were randomly assigned to a treatment of K11777 and the other five mice formed a control group injected with an equal volume of
plain water. On day 49, the researchers euthanized the mice and measured both the number of eggs and the number of worms in their livers.
Both numbers were expected to be lower in the treatment group if the drug was effective.
}
\keyword{datasets}
