% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C11.R
\docType{data}
\name{MMs}
\alias{MMs}
\title{Chapter 11 MMs}
\format{
A dataset with 55 rows and 3 columns
}
\usage{
MMs
}
\description{
This dataset shows how many brown and orange mms are counted from and mm bag.
}
\keyword{datasets}
