% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C4.R
\docType{data}
\name{Movies}
\alias{Movies}
\title{Chapter 4 Movies}
\format{
A data frame with 210 rows and four columns
}
\usage{
Movies
}
\description{
The file Movies contains the ratings and genres of movies that came out in 2008. This is simply an
observational study, as clearly movie producers do not try to create an equal number of G, PG, and R
movies each year.
}
\keyword{datasets}
