% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3.R
\docType{data}
\name{MPG}
\alias{MPG}
\title{Chapter 3 MPG Data}
\format{
A data frame with 244 rows and 12 Columns
}
\usage{
MPG
}
\description{
The MPG data compare the miles per gallon of several cars against the speed the car was going as well
as displacement. Displacement is a measure of the volume of all the cylinders within an engine. The
larger the displacement, the more quickly fuel can move through an engine, giving the vehicle more
power.
}
\keyword{datasets}
