% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Weights.R
\docType{data}
\name{Music}
\alias{Music}
\title{Music}
\format{
An object of class \code{data.frame} with 28 rows and 7 columns.
}
\usage{
Music
}
\description{
There are two Music datasets in the textbook. Please type in either MusicC1 or MusicC5 to view the correct dataset
}
\keyword{datasets}
