% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C1.R
\docType{data}
\name{MusicC1}
\alias{MusicC1}
\title{Chapter 1 Music data}
\format{
A data frame with 7 columns and 28 rows
}
\usage{
MusicC1
}
\description{
Grinnell College students Anne Tillema and Anna Tekippe conducted an experiment to study the effect
of music on a person’s level of relaxation. They hypothesized that fast songs would increase pulse
rate more than slow songs. The file called Music contains the data from their experiment.They decided to use a person’s
pulse rate as an operational definition of the person’s level of relaxation and to compare pulse rates for
two selections of music: a fast song and a slow song. For the fast song they chose “Beyond” by Nine Inch Nails,
and for the slow song they chose Rachmaninoff’s “Vocalise.” They recruited 28 student subjects for the experiment.
}
\details{
See chapter 1 page 11 for more information about this dataset.
}
\keyword{datasets}
