% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C5.R
\docType{data}
\name{MusicC5}
\alias{MusicC5}
\title{Chapter 5 Music Data}
\format{
A data frame with 160 rows and 9 columns
}
\source{
Laura Gore and Julia Lloyd, “Pop Song ‘Stayin’ Alive’ Helps People Perform Chest Compressions for CPR,” Scientific Assembly, American College of Emergency Physicians, Oct. 23, 2008, http://www.acep.org, accessed 12/2/08.
}
\usage{
MusicC5
}
\description{
Three students in an introductory statistics class read an article suggesting that listening to certain
types of music helps doctors perform chest compressions for cardiopulmonary resuscitation (CPR). It
is suggested that optimal CPR performance is 100 beats per minute. Most doctors start out at the right
pace but tend to slow down over time. Gore and Lloyd found that doctors performed much better when
listening to the song Stayin’ Alive, by the Bee Gees, which has a rhythm of 103 beats per minute.7
These students developed a research question asking if music also influences performance in
other areas. They designed an experiment testing whether music tempo or length of test (1 minute or                                                                                        3 minutes) influenced students’ ability to type fast and accurately. Would subjects listening to Stayin’ Alive type at a different speed than subjects listening to Yesterday by the Beatles?
Forty undergraduate students consented to be in the study. Each subject took four tests from
the website typingtest.com in random order based on two coin flips: 1 min/Yesterday, 1 min/Stayin’ Alive, 3 min/Yesterday, and 3 min/Stayin’ Alive. The questions the researcher wanted to test were
the effect of Song, Length, and Song*Length on words per minute (WPM)
}
\keyword{datsets}
