% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C1.R
\docType{data}
\name{NLBB_Salaries}
\alias{NLBB_Salaries}
\title{Chapter 1 NLBB Salaries}
\format{
A data frame with 15 rows and 5 columns
}
\usage{
NLBB_Salaries
}
\description{
A dataset with NLBB salaries from 2016.
}
\section{Chapter 1}{
NA
}

\keyword{datasets}
