% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C2.R
\docType{data}
\name{Normal}
\alias{Normal}
\title{Chapter 2 Normal}
\format{
A data frame with 25 rows and three columns.
}
\usage{
Normal
}
\description{
The first column is a random sample from a normal distribution. The second doubles only the 5 largest values. The third doubles only
the 5 smallest values.
}
\keyword{datasets}
