% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C7.R
\docType{data}
\name{Oscars2}
\alias{Oscars2}
\title{Chapter 7 Oscars2 Data}
\format{
a dataset with 150 rows and 7 columns
}
\usage{
Oscars2
}
\description{
While the previous information is interesting, there are some limitations in its use. The winners
of the 17 other awards are known only a few minutes before the Best Picture award is announced.
Instead of using other Academy Awards to predict the probability of Best Picture, it may be more
useful to use other award ceremonies to predict the likelihood of winning Best Picture. Hannah,
Allie, and Phillip also collected the following data:
Number of Golden Globe nominations
Number of Golden Globe wins
Number of Screen Actors Guild (SAG) nominations
Number of SAG wins
}
\keyword{datasets}
