% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C2.R
\docType{data}
\name{PaperTowel}
\alias{PaperTowel}
\title{Chapter 2 PaperTowel}
\format{
A data frame with 78 rows and three columns (Brand, Water, and Strength).
}
\usage{
PaperTowel
}
\description{
As a final project in an introductory statistics class, several students decided to conduct a study to
test the strength of paper towels. Television advertisements had claimed that a certain brand of paper
towel was the strongest, and these students wanted to determine if there really was a difference. The
students purchased rolls of towels at a local store and sampled 26 towels from 3 brands of paper towels: Bounty, Comfort, and Decorator.
Before any data were collected, these students determined that the following should be held as
constant as possible throughout the study:
\enumerate{
\item 15 drops of water were applied to the center of each towel.
\item Paper towels were selected that had the same size.
\item The towels were held at all four corners by two people.
\item Weights (10, 25, 50, or 100 grams) were slowly added to the center of each towel by a third person
until it broke.
\item The order in which the 26 paper towels were tested was randomized.
}
}
\keyword{datasets}
