% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C4.R
\docType{data}
\name{PaperTowel2}
\alias{PaperTowel2}
\title{Chapter 4 PaperTowel2}
\format{
A data frame with 234 rows and three columns
}
\usage{
PaperTowel2
}
\description{
Supplementary Data for PaperTowels.
}
\keyword{datasets}
