% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C4.R
\docType{data}
\name{PaperTowels}
\alias{PaperTowels}
\title{Chapter 4 PaperTowels}
\format{
A data frame with 156 rows and three columns
}
\usage{
PaperTowels
}
\description{
As a final project in an introductory statistics class, several students decided to conduct a study to test the
strength of paper towels. Several television advertisements had claimed that a certain brand of paper towel was
the strongest, and these students wanted to determine if there really was a difference. The students sampled
26 towels from two brands of paper towels, Comfort and Decorator.
}
\details{
Before any data were collected, these students determined that the following conditions should be held as
constant as possible throughout the study:
• Paper towels were selected that had the same size.
• The towels were held at all four corners by two people.
• Weights (10, 25, 50, 100, or 250 grams) were slowly added to the center of each towel by a third person until it broke.
In this study, there are two factors. One has two levels, Comfort (Brand C) or Decorator (Brand D), and
the other has three levels (0, 5, or 15 drops of water applied to the center of the paper towel). This leads to
2 * 3 = 6 conditions, called factor-level combinations or factorial combinations:
Brand C and 0 drops of water
Brand C and 5 drops of water
Brand C and 15 drops of water
Brand D and 0 drops of water
Brand D and 5 drops of water
Brand D and 15 drops of water
Twenty-six sheets were tested at each of the six factor-level combinations. Thus, there are 156 experimental
units used in this study. The response variable is the breaking strength of each paper towel in grams. Breaking
strength is defined as the total weight that each towel successfully held. The next additional weight caused
the towel to break.
}
\keyword{datasets}
