% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Weights.R
\docType{data}
\name{Popcorn}
\alias{Popcorn}
\title{Popcorn}
\format{
An object of class \code{character} of length 1.
}
\usage{
Popcorn
}
\description{
There are two Popcorn datasets in the textbook. Please type in either PopcornC4 or PopcornC5 to view the correct dataset
}
\keyword{datasets}
