% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C5.R
\docType{data}
\name{PopcornC5}
\alias{PopcornC5}
\title{Chapter 5 Popcorn Data}
\format{
A dataframe with 12 rows and 5 variables
}
\usage{
PopcornC5
}
\description{
Two students in a design of experiments course wanted to test if the price and the storage location of popcorn
influenced the percentage of kernels that popped.* The students purchased three boxes of both an expensive
and a generic popcorn brand (labeled Exp and Gen). Each box contained six microwavable bags of popcorn.
Two bags were randomly selected from each box and stored for one week, one in the refrigerator (Frig) and
the other at room temperature (Room). The bags were popped in random order according to the instructions
on the box, and the percentage of popped kernels was calculated for each bag.
}
\details{
*The students initially thought that temperature was the influencing factor, but a little research showed that the humidity
of the storage location was believed to influence the percentage of popped kernels. In this study, the conditions in the
refrigerator were much more humid than those in the other location.
}
\keyword{datasets}
