% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C9.R
\docType{data}
\name{Rearrest}
\alias{Rearrest}
\title{Chapter 9 Rearrest}
\format{
a dataset with 194 rows and 5 columns
}
\source{
See K. Henning and C. Frueh, “Cognitive Behavioral Treatment of Incarcerated Offenders,” Criminal Justice and Behavior, 23 (1996): 523–541.
}
\usage{
Rearrest
}
\description{
For 36 months, Henning and Frueh followed criminal activities of 194 inmates released from a medium
security prison. We can use the data from their study to investigate the time until the former inmates were
rearrested. If the former inmate had been rearrested for a criminal act before 36 months (after initial prison
release) had passed, then that former inmate’s event time is complete. If the former inmate had not been rearrested for a criminal act after 36 months had passed or had completely dropped out of the study, then that
former inmate’s event time is right censored. In addition to the time until rearrest, measurements are also
available on the following variables:
}
\details{
person: a dichotomous variable identifying former inmates who had a history of person-related
crimes—that is, those with one or more convictions for offenses such as aggravated assault or kidnapping

property: a dichotomous variable indicating whether former inmates had been convicted of a property-related crime

cenage: the “centered” age of the individual—that is, the difference between the age of the individual
on release and the average age of all inmates in the study
}
\keyword{datsets}
