% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C2.R
\docType{data}
\name{RegrTrans}
\alias{RegrTrans}
\title{Chapter 2 RegrTrans}
\format{
A data frame with 40 rows and eight columns.
}
\usage{
RegrTrans
}
\description{
Random data for the purposes of regression analysis.
}
\keyword{datasets}
