% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C1.R
\docType{data}
\name{SameMean}
\alias{SameMean}
\title{Chapter 1 SameMean dataset}
\format{
A dataset with 18 rows and 3 columns
}
\usage{
SameMean
}
\description{
This dataset is used to illuminate some key features of rank-based nonparametric tests.
}
\keyword{datasets}
