% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C7.R
\docType{data}
\name{Shuttle}
\alias{Shuttle}
\title{Chapter 7 Shuttle Data}
\format{
A data frame with 23 rows and 4 columns
}
\usage{
Shuttle
}
\description{
Table 7.1 shows the temperature at the time of each launch and whether any damage was visible in any
of the O-rings. In this chapter, we will define a successful launch as one with no evidence of any O-ring
damage. In Table 7.1, Successful Launch is a categorical variable, with 0 representing a launch where
O-ring damage occurred and 1 indicating a successful launch with no O-ring damage. Throughout the rest of
this investigation, the relatively small data set in Table 7.1 will be used to demonstrate techniques that can be
used to determine if the likelihood of O-ring damage is related to temperature.
}
\keyword{datasets}
