% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C7.R
\docType{data}
\name{SIRDS}
\alias{SIRDS}
\title{Chapter 7 SIRDS Data}
\format{
A dataset with 50 rows and two columns
}
\source{
P. K. van Vliet, and J. M. Gupta, “Sodium Bicarbonate in Idiopathic Respiratory Distress Syndrome,” Archives of Disease in Childhood, 48 (1973): 249–255.
}
\usage{
SIRDS
}
\description{
A study was conducted to determine if a relationship existed between infant birthweight and the likelihood of that infant surviving severe idiopathic respiratory distress syndrome (SIRDS). Data from
50 infants who displayed this syndrome are in the SIRDS file.
}
\keyword{datasets}
