% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C8.R
\docType{data}
\name{SkinCancer}
\alias{SkinCancer}
\title{Chapter 8 SkinCancer}
\format{
A data frame with 16 rows and five columns
}
\source{
D. Hand et al., A Handbook of Small Data Sets (London: Chapman and Hall, 1994).
}
\usage{
SkinCancer
}
\description{
The data in Table 8.4 gives the incidence of non-melanoma skin cancer among women in
Minneapolis–St. Paul by age group.
}
\keyword{datasets}
