% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C8.R
\docType{data}
\name{SkinCancer_Alternative}
\alias{SkinCancer_Alternative}
\title{Chapter 8 SkinCancer_Alternative}
\format{
A data frame with 16 rows and nine columns
}
\usage{
SkinCancer_Alternative
}
\description{
Alternative presentation of the SkinCancer Data
}
\keyword{datasets}
