% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C8.R
\docType{data}
\name{SkinCancerAlternative}
\alias{SkinCancerAlternative}
\title{Chapter 8 SkinCancerAlternative}
\format{
A data frame with 16 rows and nine columns
}
\usage{
SkinCancerAlternative
}
\description{
Alternative presentation of the SkinCancer Data
}
\keyword{datasets}
