% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C2.R
\docType{data}
\name{Skinfold}
\alias{Skinfold}
\title{Chapter 2 Skinfold}
\format{
A data frame with 29 rows and three columns (Disease, Thickness, and ln. Thickness).
}
\usage{
Skinfold
}
\description{
Celiac disease results in an inability to absorb carbohydrates and fats. Crohn’s disease is another
chronic intestinal disease in which the body’s immune system attacks the intestines. Both Crohn’s
disease and celiac disease often result in malnutrition or impaired growth in children. A skinfold
thickness measurement is a simple technique assessing body fat percentages by pinching the skin
near the biceps and then using a calipers to measure the skin thickness.
}
\keyword{datasets}
