% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C8.R
\docType{data}
\name{Smoking}
\alias{Smoking}
\title{Chapter 8 Smoking}
\format{
A data frame with one row and seven columns
}
\usage{
Smoking
}
\description{
Several regular smokers were selected from a larger Minnesota survey about smoking habits. All reported
smoking about the same number of cigarettes per day in recent months.
Each was asked to report on the number of cigarettes smoked the day before (January 26) during the same 2-hour period. They were also asked
whether they had been at work or at home or elsewhere during that 2-hour period.
}
\keyword{datasets}
