% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C4.R
\docType{data}
\name{Soda}
\alias{Soda}
\title{Chapter 4 Soda}
\format{
A data frame with 24 rows and four columns
}
\usage{
Soda
}
\description{
Two students, Julie and Daphne, conducted an experiment to test the effects of soda type (Pepsi
vs. 7-Up), angle of the cup (cup flat on the table or slightly tipped), and soda temperature (refrigerated
at 5°C vs. room temperature at 21°C) on the height of fizz produced when soda was poured out of a
can into a cup. For each of the 24 trials, these students poured a can of either Pepsi or 7-Up into a clear
cup and measured the peak fizz (in centimeters) produced, using a ruler on the outside of the cup.
}
\keyword{datasets}
