% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stats2Labs.R
\docType{package}
\name{Stats2Labs}
\alias{Stats2Labs}
\title{Stats2Labs Package}
\description{
This package contains all the datasets in the textbook, Practicing Statistics, by Shonda Kuiper and Robert Sklar.
To call a dataset, simply type the name as it appears in the textbook. If there are spaces in the dataset, use an
underscore (_) between the two words (EX: NLBB_Salaries). Click the index button to see all the datasets in the package.
}
