% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C10.R
\docType{data}
\name{Stock2010}
\alias{Stock2010}
\title{Chapter 10 Stock2010}
\format{
A data frame with 252 rows and 26 columns
}
\usage{
Stock2010
}
\description{
The data set 2010Stock contains 2010 daily closing stock market values for several well-known
companies.
}
\keyword{datasets}
