% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C10.R
\docType{data}
\name{Stocks2006}
\alias{Stocks2006}
\title{Chapter 10 Stocks2006}
\format{
A data frame with 251 rows and four columns
}
\usage{
Stocks2006
}
\description{
The stock market data contains dates for each business day in 2006 for the three major indexes.
}
\keyword{datasets}
