% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C10.R
\docType{data}
\name{Stocks_2006}
\alias{Stocks_2006}
\title{Chapter 10 Stocks_2006}
\format{
A data frame with 251 rows and four columns
}
\usage{
Stocks_2006
}
\description{
The stock market data set also contains dates for each business day in 2006.
}
\keyword{datasets}
