% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C6.R
\docType{data}
\name{Table6.1}
\alias{Table6.1}
\title{Chapter 6 Table6.1}
\format{
A data frame with two rows and three columns
}
\usage{
Table6.1
}
\description{
Table 6.1 shows data from 37 FNA slide samples. Slides with smooth ellipsoid-shaped nuclei were classified
as round, and slides with poorly shaped cell nuclei were classified as concave. A biopsy was also conducted
on each of these samples to determine if each was malignant or benign.
}
\keyword{datasets}
