% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C6.R
\docType{data}
\name{Table6.10}
\alias{Table6.10}
\title{Chapter 6 Table 6.10}
\format{
A table with three rows and three columns
}
\source{
D. Salsburg, The Lady Tasting Tea (New York: Freeman, 2001)
}
\usage{
Table6.10
}
\description{
At a summer tea party in the late 1920s, a lady claimed that she could determine by taste if tea had
been poured into the milk or if milk had been poured into the tea. Sir Ronald Fisher, well known
for his work in design of experiments, discussed how to test this lady’s claim. The lady was asked
to taste several cups in random order and identify each as tea first or milk first. Although the actual
results were not published, rumor has it that the lady correctly identified every single cup.
Let’s assume that the lady was given 14 cups of tea in random order. She was told that half the
cups had tea first and the other half had milk first. Table 6.10 provides hypothetical results.
}
\keyword{datasets}
