% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C6.R
\docType{data}
\name{Table6.11}
\alias{Table6.11}
\title{Chapter 6 Table 6.11}
\format{
A table with three rows and four columns
}
\usage{
Table6.11
}
\description{
Members of the Donner party attempted a new route between Fort Bridger, Wyoming, and the Humboldt
River, Nevada. This new route took much longer than expected, and the entire party was trapped in the
Sierra Nevada in the winter of 1846–1847. By the time they were rescued in the spring, many members
had died. Table 6.11 lists the gender and survival status of all adults who were trapped during this trip.
}
\keyword{datasets}
