% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C6.R
\docType{data}
\name{Table6.14}
\alias{Table6.14}
\title{Chapter 6 Table 6.14}
\format{
A data frame with three rows and four columns
}
\usage{
Table6.14
}
\description{
The extended activities in Section 6.7 discussed a clinical study on Zocor, a drug used in lowering
cholesterol, to determine if there was a difference between the treatment (Zocor) and the placebo in
reducing the number of deaths from heart attacks. Table 6.14 shows the original data from this study.
}
\keyword{datasets}
