% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C6.R
\docType{data}
\name{Table6.5}
\alias{Table6.5}
\title{Chapter 6 Table 6.5}
\format{
A data frame with two rows and three columns
}
\usage{
Table6.5
}
\description{
In a retrospective study of lung cancer patients in 20 London hospitals,
Richard Doll identified a relationship between smoking and lung cancer.
Table 6.5 shows a partial set of data from his study. The data were
collected on 60 female patients with lung cancer and 60 control females.
}
\keyword{datasets}
