% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C7.R
\docType{data}
\name{Table7.3}
\alias{Table7.3}
\title{Chapter 7 Table 7.3 data}
\format{
A table with 5 rows and 5 columns
}
\usage{
Table7.3
}
\description{
Table 7.3 shows the cancer cell data with the Radius variable grouped into five levels. Clearly grouping is
not necessary for logistic regression, but this grouping is done to demonstrate how to conduct logistic regression when the response is binomial.
}
\keyword{datasets}
