% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C9.R
\docType{data}
\name{Table9.2}
\alias{Table9.2}
\title{Chapter 9 Table 9.2}
\format{
A table with 7 rows and 2 columns
}
\usage{
Table9.2
}
\description{
This dataset is another chocolate chip melting time data.
}
\keyword{datasets}
