% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C6.R
\docType{data}
\name{Table_6.12}
\alias{Table_6.12}
\title{Chapter 6 Table_6.12}
\format{
A data frame with four rows and four columns
}
\usage{
Table_6.12
}
\description{
Many people with cocaine addictions also suffer from mood disorders like depression or a manicdepressive disorder.
To aid in the process of trying to break a patient’s addiction to cocaine, a
common antidepressant known as lithium is often given to patients in drug treatment centers. A
randomized, comparative experiment was performed to determine whether another commonly
prescribed antidepressant, disipramine, could be used to reduce the likelihood of patients’ using
cocaine again or whether antidepressant therapy helped at all. The 72 patients selected for study
over a 3-year period were randomly assigned to one of three treatments: disipramine, lithium, or
a placebo. After the end of the 6-week treatment period, these patients were interviewed to determine if they
had been successful in stopping their cocaine habit. A success was defined as having
stopped using cocaine for 3 weeks or more during the 6-week treatment period. The data from the
study are shown in Table 6.12.
}
\keyword{datasets}
