% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C6.R
\docType{data}
\name{Table_6.13}
\alias{Table_6.13}
\title{Chapter 6 Table_6.13}
\format{
A data frame with one row and eight columns
}
\usage{
Table_6.13
}
\description{
A local organization conducted a traffic study to determine if traffic collisions were more likely on
Friday and Saturday nights than on other nights. The city council was considering a proposal that
bars in a specific district be closed earlier in an effort to reduce late-night accidents in the area. The
assumption was that a larger number of traffic accidents would correspond to the heaviest bar traffic
(which occurred on the weekends). Table 6.13 shows data collected on traffic accidents that had
occurred in that district over the past two years.
}
\keyword{datasets}
