% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C6.R
\docType{data}
\name{Table_6.15}
\alias{Table_6.15}
\title{Chapter 6 Table_6.15}
\format{
A data frame with four rows and four columns
}
\usage{
Table_6.15
}
\description{
In October 1995, the United Kingdom Committee on Safety of Medicines (CSM) issued a warning
to 190,000 general practitioners, pharmacists, and directors of public health about oral contraceptive
pills containing gestodene or desogestrel. The warning, based on three unpublished epidemiological
research studies, stated, “It is well known that the pill may rarely produce thrombosis (blood clots)
involving veins of the legs. New evidence has become available indicating that the chance of thrombosis
occurring in a vein increases about two-fold for some types of pills compared to others.
Table 6.15 provides data from one of the studies.
}
\keyword{datasets}
