% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C6.R
\docType{data}
\name{Table_6.7}
\alias{Table_6.7}
\title{Chapter 6 Table_6.7}
\format{
A data frame with six rows and two columns
}
\usage{
Table_6.7
}
\description{
Your friend offers to play a simple game with you.
He has a six-sided die and will roll it 30 times. Each time
he rolls a 5 or 6, you will pay him $5. Each time he rolls
a 1, 2, 3, or 4, he will pay you $5. The results of your
friend’s 30 rolls for this game are shown in Table 6.7.
}
\keyword{datasets}
