% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C6.R
\docType{data}
\name{Table_6.9}
\alias{Table_6.9}
\title{Chapter 6 Table_6.9}
\format{
A data frame with two rows and three columns
}
\usage{
Table_6.9
}
\description{
A study was conducted by introductory statistics students to see whether majors in the sciences are
as likely to take a statistics course as majors in the humanities. They sampled 50 seniors from among
humanities majors and another 50 seniors from among science majors and asked if they had taken a
statistics course or were scheduled to take one before they graduated. Their data are provided in Table 6.9.
}
\keyword{datasets}
