% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C8.R
\docType{data}
\name{Table_8.1}
\alias{Table_8.1}
\title{Chapter 8 Table_8.1}
\format{
A data frame with eight rows and six columns
}
\usage{
Table_8.1
}
\description{
Cancer cases and cancer incidence rates (CIR) by age group for the Bartlett–Green
Acres neighborhood of Randolph (BGA), 1964–1984, and similar data from the
Connecticut Tumor Registry (CTR), 1980–1984.
}
\keyword{datasets}
