% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C7.R
\docType{data}
\name{Tattoos}
\alias{Tattoos}
\title{Chapter 7 Tattoos Data}
\format{
A dataset with 55 rows and 4 columns
}
\source{
. Data are modified from A. D. Lunn and D. R. McNeil, The SPIDA Manual (Sydney: Statistical Computing Laboratory, 1988).
}
\usage{
Tattoos
}
\description{
Lunn and McNeil show a data set comparing two methods of surgical tattoo removal. In addition to
which method was used, the depth of the tattoo and the patient’s gender were recorded. In this data
set, Removal = 1 represents a successful removal and Removal = 0 represents a poor removal.
}
\keyword{datasets}
