% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C5.R
\docType{data}
\name{Tennis}
\alias{Tennis}
\title{Chapter 5 Tennis Data}
\format{
A data frame with 150 rows and 5 columns
}
\source{
Rob Bower and Rod Cross, “String Tension Effects on Tennis Ball Rebound Speed and Accuracy During Playing Conditions,” Journal of Sports Sciences, 23.7 (July 2005): 765–771.
}
\usage{
Tennis
}
\description{
Students on the college tennis team were interested in knowing if string tension affected the speed
and accuracy of a tennis ball. They had recently read an article in the Journal of Sports Sciences
and decided conduct a similar study for themselves.6
After warming up, five men’s varsity tennis
players volunteered to hit 30 serves using three different racquet tensions over three days. Day was
not treated as a factor of interest in the study, but used to avoid fatigue in the players. For each serve,
players aimed at a target in the back center of the service box, and accuracy was measured as the
distance in inches from the center of the box to the ball strike location. A radar gun was used to measure the velocity of each player’s serve.
}
\keyword{datasets}
