% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C10.R
\docType{data}
\name{Test}
\alias{Test}
\title{Chapter 10 Test}
\format{
A data frame with 113 rows and four columns
}
\usage{
Test
}
\description{
The file Test contains several standardized variables.
}
\keyword{datasets}
