% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C10.R
\docType{data}
\name{TestC10}
\alias{TestC10}
\title{Chapter 10 TestC10}
\format{
A data frame with 113 rows and four columns
}
\usage{
TestC10
}
\description{
The file Test contains several standardized variables.
}
\keyword{datasets}
