% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3.R
\docType{data}
\name{`2008Movies`}
\alias{`2008Movies`}
\alias{2008Movies}
\title{Chapter 3 2008Movies Data}
\format{
A data frame with 199 rows and 12 columns
}
\usage{
`2008Movies`
}
\description{
The 2008Movies file contains data on movies released in 2008
}
\keyword{datasets}
