% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3.R
\docType{data}
\name{`4-8Cyl`}
\alias{`4-8Cyl`}
\alias{4-8Cyl}
\title{Chapter 3 4-8Cyl Data}
\format{
A data frame with 244 rows and 12 Columns
}
\usage{
`4-8Cyl`
}
\description{
The data set is a subset of the Cars dataset that only contains cars with either 4 or 8 cyl. Type ?Cars for more information.
}
\keyword{datasets}
