% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C2.R
\docType{data}
\name{Tires}
\alias{Tires}
\title{Chapter 2 Tires}
\format{
A data frame with 30 rows and two columns (Brand and Miles).
}
\usage{
Tires
}
\description{
The data file Tires relates to five brands of tires chosen at random from local stores. Six tires of each
brand were selected and placed in random order on a machine that tested tread durability in thousands
of miles.
}
\keyword{datasets}
