% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C4.R
\docType{data}
\name{Towels2}
\alias{Towels2}
\title{Chapter 4 Towels2}
\format{
A data frame with 234 rows and three columns
}
\usage{
Towels2
}
\description{
Supplementary Data for PaperTowels which include the brand Bounty.
}
\keyword{datasets}
