% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C10.R
\docType{data}
\name{Turtles}
\alias{Turtles}
\title{Chapter 10 Turtles}
\format{
A data frame with 48 rows and four columns
}
\source{
P. Jolicoeur and J. E. Mosimann, “Size and Shape Variation in the Painted Turtle: A Principal Component Analysis,” Growth, 24 (1960): 339–354.
}
\usage{
Turtles
}
\description{
Jolicoeur and Mosimann measured the length, width, and height of 48 painted
turtle shells and conducted PCA on the data. Their work has been influential
in the field of allometry, the study of the
relative growth of a part of an organism in relation to an entire organism.
}
\keyword{datasets}
