% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Twins.R
\docType{data}
\name{Twins}
\alias{Twins}
\title{Chapter 1 Twins dataset}
\format{
A dataset with 15 rows and 4 columns
}
\usage{
Twins
}
\description{
In a 1990 study by Suddath et al., reported in Ramsey and Schafer,12 researchers used magnetic resonance
imaging to measure the volume of various regions of the brain for a sample of 15 monozygotic twins, where
one twin was affected with schizophrenia and the other was unaffected. The twins were from North America and
comprised eight male pairs, and seven female pairs ranging in age from 25 to 44 at the time of the study.
The sizes in volume (cm3) of the hippocampus are in the file called Twins.
}
\keyword{datasets}
