% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C1.R
\docType{data}
\name{Variable_DescriptionsC1}
\alias{Variable_DescriptionsC1}
\title{An index of the Chapter 1 Data included in this package.}
\format{
A data frame with 47 rows and four columns (Chapter, Data.Set, Variable.Name, and Description).
}
\usage{
Variable_DescriptionsC1
}
\description{
An index of the Chapter 1 Data included in this package.
}
\keyword{datasets}
