% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3.R
\docType{data}
\name{Variable_DescriptionsC3}
\alias{Variable_DescriptionsC3}
\title{Chapter 3 Variable_Descriptions}
\format{
A data frame with 96 rows and four columns (Chapter, Data.Set, Variable.Name, and Description).
}
\usage{
Variable_DescriptionsC3
}
\description{
An index of the Chapter 2 Data included in this package.
}
\keyword{datasets}
