% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C1.R
\docType{data}
\name{VariableDescriptionsC1}
\alias{VariableDescriptionsC1}
\title{Chapter 1 Variable Descriptions}
\format{
A data frame with 47 rows and four columns (Chapter, Data.Set, Variable.Name, and Description).
}
\usage{
VariableDescriptionsC1
}
\description{
Chapter 1 Variable Descriptions
}
\keyword{datasets}
