% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C10.R
\docType{data}
\name{VariableDescriptionsC10}
\alias{VariableDescriptionsC10}
\title{Chapter 10 VariableDescriptionsC10}
\format{
A data frame with 114 rows and 16 columns
}
\usage{
VariableDescriptionsC10
}
\description{
A set of Chapter 10 data set descriptions.
}
\keyword{datasets}
