% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C11.R
\docType{data}
\name{VariableDescriptionsC11}
\alias{VariableDescriptionsC11}
\title{Chapter 11 VariableDescriptionsC11}
\format{
An object of class \code{data.frame} with 3 rows and 4 columns.
}
\usage{
VariableDescriptionsC11
}
\description{
This dataset contains descriptions of all the variables of datasets within chapter 11.
}
\keyword{datasets}
