% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C2.R
\docType{data}
\name{VariableDescriptionsC2}
\alias{VariableDescriptionsC2}
\title{Chapter 2 Variable Descriptions}
\format{
A data frame with 35 rows and four columns (Chapter, Data.Set, Variable.Name, and Description).
}
\usage{
VariableDescriptionsC2
}
\description{
An index of the Chapter 2 Data included in this package.
}
\keyword{datasets}
