% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C4.R
\docType{data}
\name{VariableDescriptionsC4}
\alias{VariableDescriptionsC4}
\title{Chapter 4 Variable Descriptions}
\format{
A data frame with 43 rows and four columns
}
\usage{
VariableDescriptionsC4
}
\description{
An index of the chapter 4 data included in this package.
}
\keyword{datasets}
