% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C5.R
\docType{data}
\name{VariableDescriptionsC5}
\alias{VariableDescriptionsC5}
\title{Chapter 5 Variable Descriptions}
\format{
An object of class \code{data.frame} with 65 rows and 4 columns.
}
\usage{
VariableDescriptionsC5
}
\description{
All the variable descriptions for the datasets in Chapter 5.
}
\keyword{datasets}
