% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C7.R
\docType{data}
\name{VariableDescriptionsC7}
\alias{VariableDescriptionsC7}
\title{Chapter 7 Variable Descriptions}
\format{
a dataset with 65 rows and 4 columns
}
\usage{
VariableDescriptionsC7
}
\description{
Contains all the variable descriptions for the datasets in chapter 7.
}
\keyword{datasets}
