% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C9.R
\docType{data}
\name{VariableDescriptionsC9}
\alias{VariableDescriptionsC9}
\title{Chapter 9 Variable Descriptions}
\format{
An object of class \code{data.frame} with 30 rows and 4 columns.
}
\usage{
VariableDescriptionsC9
}
\description{
This dataset has variable descriptions for all the datasets in chapter 9
}
\keyword{datasets}
