% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C10.R
\docType{data}
\name{Versicolor}
\alias{Versicolor}
\title{Chapter 10 Versicolor}
\format{
A data frame with 50 rows and four columns
}
\usage{
Versicolor
}
\description{
The data set Versicolor contains four measurements (sepal length,
sepal width, petal length, and petal width) on 50 versicolor irises.
The petals are colorful and flashy,
while the sepal (under the petals) tends to be less colorful.
}
\keyword{datasets}
