% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Weights.R
\docType{data}
\name{Weights}
\alias{Weights}
\title{Weights}
\format{
An object of class \code{character} of length 1.
}
\usage{
Weights
}
\description{
There are two Weights datasets in the textbook. Please type in either C2Weights or C3Weights to view the correct dataset
}
\keyword{datasets}
