% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3.R
\docType{data}
\name{WeightsC3}
\alias{WeightsC3}
\title{Chapter 3 Weights Data}
\format{
A data frame with 62 rows and 3 Columns
}
\source{
T. Allison and D. Cicchetti, “Sleep in Mammals: Ecological and Constitutional Correlates,” Science, 194 (Nov. 12,
1976): 732–734.
}
\usage{
WeightsC3
}
\description{
In the process of studying the effect of sleep on mammals, T. Allison and D. Cicchetti collected the
brain weights (in grams) and body weights (in kilograms) of 62 mammals.
In the data set, you can see the measurements for these two variables for each of the 62 animals studied.
}
\keyword{datasets}
