% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C2.R
\docType{data}
\name{Winglength1}
\alias{Winglength1}
\title{Chapter 2 Winglength1}
\format{
A data frame with 15 rows and three columns (Experiment, WingLength, and FlightTime).
}
\usage{
Winglength1
}
\description{
Data concerning a helicopter research project measuring the optimal winglength.
}
\keyword{datasets}
