% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C2.R
\docType{data}
\name{Winglength2}
\alias{Winglength2}
\title{Chapter 2 Winglength2}
\format{
A data frame with 30 rows and three columns (WingLength, FlightTime, and Order).
}
\usage{
Winglength2
}
\description{
Additional data concerning a helicopter research project measuring the optimal winglength.
}
\keyword{datasets}
