% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C7.R
\docType{data}
\name{Cancer2}
\alias{Cancer2}
\title{Chapter 7 Cancer2 data}
\format{
A data frame with 569 observations with 3 columns
}
\source{
W. Wolberg and O. Mangasarian, “Multisurface Method of Pattern Separation for Medical Diagnosis Applied to Breast Cytology,” Proceedings of the National Academy of Sciences of the United States of America, 87. 23 (Dec. 1990): 9193–9196.
}
\usage{
Cancer2
}
\description{
In this data set, radius is actually the average radius (in micrometers, mm) of all visible cell nuclei from a slide, but we will refer to this variable simply as the cell radius for the tumor. The concavity
of the cell nuclei is an indicator of whether the visible cell nuclei from the sample have the nice round shape
of typical healthy cells or whether cells appear to have grown in such a way that the perimeters of the cell
nuclei tend to have concave points. For more information about how the data was collected, see page 224.
}
\keyword{datasets}
