% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C2.R
\docType{data}
\name{Games2}
\alias{Games2}
\title{Chapter 2 Games2}
\format{
A data frame with 40 rows and six columns (studentID, numpegs, Type, Time, Hand, and Type2).
}
\usage{
Games2
}
\description{
An extension on the Games1 dataset
}
\keyword{datasets}
