% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C2.R
\docType{data}
\name{Games2}
\alias{Games2}
\title{Chapter 2 Games2}
\format{
A data frame with 40 rows and six columns (studentID, numpegs, Type, Time, Hand, and Type2).
}
\usage{
Games2
}
\description{
Several students in an introductory statistics class wanted to develop a final project that would test the
impact of distracters. They decided to conduct a study to determine if students at their college would perform
differently when a distracting color was incorporated into a computerized game. This game challenges people
to place an assortment of shaped pegs into the appropriate spaces as quickly as possible.
Before any data were collected, these students developed a clear set of procedures:
\enumerate{
\item 40 students would be randomly selected from the college.
\item 20 students would be assigned to the standard game and 20 would be assigned to a game with a color
distracter. The student researchers would flip a coin to randomly assign subjects to a treatment. Once
20 subjects had been assigned to either group, the rest would automatically be assigned to play the
other game.
\item Subjects would see a picture of the game and have the rules clearly explained to them before they
played the game.
\item Subjects would play the game in the same area with similar background noise to control for other
possible distractions.
\item The response variable would be the time in seconds from when the participant pressed the “start
game” button to when he or she won the game.
}
}
\details{
Additionally, they were interested in testing whether college
students could play the game more quickly with their right or left hand.
The data set Games2 shows a column Type2 with four types of games, based on distracter
and which hand was used.
}
\keyword{datasets}
