% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C7.R
\docType{data}
\name{Oscars}
\alias{Oscars}
\title{Chapter 7 Oscars Data}
\format{
A dataset with 48 rows and 20 columns
}
\usage{
Oscars
}
\description{
In 2009, three Grinnell students (Allie Greenberg, Hannah Lytle, and Phillip Brogdon) conducted an
analysis to estimate the probability of winning the Academy Award for Best Picture. The Academy
Awards, or “Oscars,” are given annually to honor high achievement in the film industry. The Academy consists of over 6000 members who nominate their colleagues and vote to decide on the winners
of this prestigious award.
}
\keyword{datasets}
