% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3.R
\docType{data}
\name{Politics}
\alias{Politics}
\title{Chapter 3 Politics Data}
\format{
A data frame with 50 rows and 3 column (Column names are not conventional)
}
\usage{
Politics
}
\description{
In 2001, several college students collected data from 50 states on each of these factors from the U.S. Census Bureau.
}
\keyword{datasets}
