% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C4.R, R/C5.R
\docType{data}
\name{Popcorn}
\alias{Popcorn}
\title{Chapter 4 Popcorn}
\format{
A data frame with 32 rows and five columns

A dataframe with 12 rows and 5 variables
}
\usage{
Popcorn

Popcorn
}
\description{
Two popcorn lovers, who also happened to be taking a statistics course, decided to test whether there is a
difference in quality between microwave popcorn brands. Yvonne and Tue wanted to know if a cheaper brand
of popcorn was just as good as more expensive popcorn. These students could have chosen to conduct a study
that could be analyzed with a two-sample t-test if they had simply compared two brands of popcorn. However,
if they did a two-sample t-test, they would need to hold many factors constant, such as the type of microwave,
cooking time, and storage procedures. Since Yvonne and Tue believed that some of these factors could also
impact the quality of the popcorn, they decided to include some of these additional factors in their study.

Two students in a design of experiments course wanted to test if the price and the storage location of popcorn
influenced the percentage of kernels that popped.* The students purchased three boxes of both an expensive
and a generic popcorn brand (labeled Exp and Gen). Each box contained six microwavable bags of popcorn.
Two bags were randomly selected from each box and stored for one week, one in the refrigerator (Frig) and
the other at room temperature (Room). The bags were popped in random order according to the instructions
on the box, and the percentage of popped kernels was calculated for each bag.
}
\keyword{datasets}
