% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C9.R
\docType{data}
\name{Veteran}
\alias{Veteran}
\title{Chapter 9 Veteran}
\format{
A dataset with 137 rows and 8 columns
}
\source{
D. Kalbfleisch and R. L. Prentice, The Statistical Analysis of Failure Time Data, 2nd ed. (New York: Wiley, 2002).
}
\usage{
Veteran
}
\description{
The U.S. Veterans Administration lung cancer trial studied 137 males with advanced inoperable lung
cancer. The subjects were randomly assigned to either a standard chemotherapy treatment or a test
chemotherapy treatment. Several additional variables were also measured on the subjects. The data in
the file Veteran have been investigated in Kalbfleisch and Prentice. See page 326 for brief descriptions of
the variables.
}
\keyword{datasets}
