% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Weights.R
\docType{data}
\name{Weights}
\alias{Weights}
\title{Weights}
\format{
An object of class \code{character} of length 1.
}
\usage{
Weights
}
\description{
Type in WeightsC2 instead of Weight if you are looking at problem 32 in chapter 2. Otherwise, there are two datasets of weights,
WeightsC2 and WeightsC3
}
\keyword{datasets}
