% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C2.R
\docType{data}
\name{Weights}
\alias{Weights}
\title{Chapter 2 Weights}
\format{
A data frame with 30 rows and six columns (species, bodyweight, brainweight, maximumlifespan, gestationtime, and totalsleep).
}
\usage{
Weights
}
\description{
In addition to brain weights and body weights, the Weight data set contains information on lifespan,
gestation, and hours of sleep per day.
}
\keyword{datasets}
